+++ src/components/TodoList.tsx
@@ -1,3 +1,5 @@
+import { useContext } from 'react';
+import TodoContext from '../store/todo-context';
 import { Todo } from '../types/todo';
 import TodoItem from './TodoItem';
 interface TodoListInterface {
@@ -6,6 +8,8 @@ interface TodoListInterface {
 }
 
 const TodoList = ({ todos, removeTodo }: TodoListInterface) => {
+	const todoCtx = useContext(TodoContext);
+
 	return (
 		<ul>
 			{todos.map(item => (
