+++ src/store/todo-context.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+const TodoContext = createContext({
+	todos: [],
+	addTodo: (text: string) => {},
+	removeTodo: (id: string) => {},
+});
+
+export default TodoContext;
